ADFA-11587: Fix TreeSitter native library loading during tests to prevent failures
