+++ app/src/main/java/com/itsaky/androidide/fragments/output/LogViewFragment.kt
@@ -32,6 +32,7 @@ import com.itsaky.androidide.fragments.EmptyStateFragment
 import com.itsaky.androidide.models.LogLine
 import com.itsaky.androidide.utils.ILogger.Level
 import com.itsaky.androidide.utils.jetbrainsMono
+import com.itsaky.androidide.utils.isTestMode
 import io.github.rosemoe.sora.widget.style.CursorAnimator
 import org.slf4j.LoggerFactory
 import java.util.concurrent.ArrayBlockingQueue
@@ -258,26 +259,6 @@ abstract class LogViewFragment :
     }
   }
 
-  private fun isTestMode(): Boolean {
-    // Check system property (for unit tests)
-    if (System.getProperty("androidide.test.mode") == "true") {
-      return true
-    }
-    
-    // Check if we're running under test instrumentation
-    try {
-      val instrumentationClass = Class.forName("androidx.test.platform.app.InstrumentationRegistry")
-      val getInstrumentationMethod = instrumentationClass.getMethod("getInstrumentation")
-      val instrumentation = getInstrumentationMethod.invoke(null)
-      if (instrumentation != null) {
-        return true
-      }
-    } catch (e: Exception) {
-      // Not running under instrumentation
-    }
-    
-    return false
-  }
 
   override fun onDestroyView() {
     _binding?.editor?.release()
